# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/admin/statistics": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Statistics",
        "description": "Gets the statistics of the admin.",
        "operationId": "get_admin_statistics",
        "parameters": [
          {
            "name": "game_mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Mode"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/texts": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Texts",
        "description": "Gets all texts.",
        "operationId": "get_texts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextWithStatistics"
                  },
                  "title": "Response Get Texts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/texts/{text_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Text",
        "description": "Gets a text by the given id.",
        "operationId": "get_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextWithQuestionsAndStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Text",
        "description": "Deletes a text by the given id.",
        "operationId": "delete_text",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Text"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/texts/{text_id}/questions": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Questions",
        "description": "Gets the questions of a text by the given id.",
        "operationId": "get_questions",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionWithStatistics"
                  },
                  "title": "Response Get Questions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/texts/{text_id}/questions/{question_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Question",
        "description": "Gets a question of a text by the given id.",
        "operationId": "get_question",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionWithStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Question",
        "description": "Deletes a question of a text by the given id.",
        "operationId": "delete_question",
        "parameters": [
          {
            "name": "text_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/generate-text": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Generate Text",
        "operationId": "generate_text",
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Difficulty"
            }
          },
          {
            "name": "fiction",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Fiction"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCreateWithQuestions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/approve-text": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Approve Text",
        "description": "Adds a text to the database.",
        "operationId": "approve_text",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCreateWithQuestions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextWithQuestions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminStatistics": {
        "properties": {
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          }
        },
        "type": "object",
        "required": [
          "average_score",
          "average_wpm",
          "min_wpm",
          "max_wpm"
        ],
        "title": "AdminStatistics"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_option": {
            "type": "integer",
            "title": "Correct Option"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "content",
          "options",
          "correct_option",
          "id"
        ],
        "title": "Question"
      },
...
      "Text": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id"
        ],
        "title": "Text"
      },
...
      "TextWithQuestionsAndStatistics": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id",
          "questions",
          "min_wpm",
          "max_wpm",
          "average_wpm",
          "average_score"
        ],
        "title": "TextWithQuestionsAndStatistics"
      },
      "TextWithStatistics": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fiction": {
            "type": "boolean",
            "title": "Fiction"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "min_wpm": {
            "type": "integer",
            "title": "Min Wpm"
          },
          "max_wpm": {
            "type": "integer",
            "title": "Max Wpm"
          },
          "average_wpm": {
            "type": "integer",
            "title": "Average Wpm"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "summary",
          "source",
          "fiction",
          "difficulty",
          "word_count",
          "description",
          "author",
          "image_url",
          "id",
          "min_wpm",
          "max_wpm",
          "average_wpm",
          "average_score"
        ],
        "title": "TextWithStatistics"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}